Remove helper modules
