--- packages/ciphernode/core/src/plaintext_writer/mod.rs
@@ -1,2 +1,0 @@
-mod plaintext_writer;
-pub use plaintext_writer::*;
