--- packages/ciphernode/core/src/public_key_writer/mod.rs
@@ -1,2 +1,0 @@
-mod public_key_writer;
-pub use public_key_writer::*;
